+++ bld/clib/_dos/c/allocos2.c
@@ -24,8 +24,7 @@
 *
 *  ========================================================================
 *
-* Description:  WHEN YOU FIGURE OUT WHAT THIS FILE DOES, PLEASE
-*               DESCRIBE IT HERE!
+* Description:  Emulate _dos_allocmem()/_dos_freemem() on OS/2.
 *
 ****************************************************************************/
 
@@ -42,7 +41,7 @@
 #if defined(__386__) || defined(__PPC__)
   typedef void                  *mem_id;
 #elif defined( _M_I86 )
-  typedef unsigned short        mem_id;
+  typedef unsigned              mem_id;
   #if defined(__BIG_DATA__)
     #define MODIFIES ds es
   #else
